Add HasTokenValue trait and remove tokenValue method from HasTemplate trait.
