+++ src/Attribute/Custom/HasTemplate.php
@@ -14,7 +14,6 @@ use function strtr;
 trait HasTemplate
 {
     protected string $template = '';
-    protected array $tokenValue = [];
 
     /**
      * Set the template.
@@ -27,14 +26,6 @@ trait HasTemplate
     {
         $new = clone $this;
         $new->template = $value;
-
-        return $new;
-    }
-
-    public function tokenValue(string $token, string $value): static
-    {
-        $new = clone $this;
-        $new->tokenValue[$token] = $value;
 
         return $new;
     }
