+++ src/Base/AbstractElement.php
@@ -17,6 +17,7 @@ abstract class AbstractElement extends Element
     use Attribute\Custom\HasContent;
     use Attribute\Custom\HasPrefixAndSuffix;
     use Attribute\Custom\HasTemplate;
+    use Attribute\Custom\HasTokenValue;
     use Attribute\HasClass;
     use Attribute\HasData;
     use Attribute\HasId;
