+++ tests/Attribute/Custom/HasTokenValueTest.php
@@ -4,17 +4,17 @@ declare(strict_types=1);
 
 namespace PHPForge\Html\Tests\Attribute\Custom;
 
-use PHPForge\Html\Attribute\Custom\HasTemplate;
+use PHPForge\Html\Attribute\Custom\HasTokenValue;
 use PHPUnit\Framework\TestCase;
 
-final class HasTemplateTest extends TestCase
+final class HasTokenValueTest extends TestCase
 {
     public function testImmutability(): void
     {
         $instance = new class () {
-            use HasTemplate;
+            use HasTokenValue;
         };
 
-        $this->assertNotSame($instance, $instance->template(''));
+        $this->assertNotSame($instance, $instance->tokenValue('', ''));
     }
 }
